partial refactor to change namespace to Metarma, and also to have a station class
Main refactor here is to take some of the logic and binding that was in
Metar and make is richer to implement the Observer pattern so I can get
some MVC love.

Created Station and StationList classes. Some of the responsibilites of
Metar and the view will now be moved into Station. The view will just
observe and have very little logic.