refactor Metarma to use bindings and DataTemplates for the view
Big problem now is that attempting to run the tasks asynchonously is not
updating the UI. Everything works OK if I wait, but that's not really
what I want. I've struggled with this when I was originally building it.
If I just used async/await I would end up with deadlocks, using the
Task.Run means things are executed on the wrong thread. I think.